backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
ratelimit==2.2.1
setuptools==80.7.1
sqlparse==0.5.3
typing_extensions==4.13.1
